pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "octopus-export-optimizer"
version = "0.1.0"
description = "Export optimisation engine for Octopus Energy Agile tariff with battery storage"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "batesyx" },
]

dependencies = [
    "httpx>=0.27,<1.0",
    "pydantic>=2.5,<3.0",
    "pydantic-settings>=2.1,<3.0",
    "paho-mqtt>=2.0,<3.0",
    "apscheduler>=3.10,<4.0",
    "python-dotenv>=1.0,<2.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "respx>=0.21,<1.0",
    "time-machine>=2.13,<3.0",
    "ruff>=0.3,<1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/octopus_export_optimizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
